Refactor tool editor, add mask orientation toggles
